project/requirements.txt
pandas==2.0.0
PyYAML==6.0